+++ src/app/components/ServiceWorkerRegister/ServiceWorkerRegisterCatalog.tsx
@@ -1,0 +1,10 @@
+'use client'
+
+import { ServiceWorkerRegister } from './ServiceWorkerRegister'
+
+export const ServiceWorkerRegisterCatalog = () => (
+  <div className="p-6">
+    <ServiceWorkerRegister />
+    <span>Service worker registration logic</span>
+  </div>
+)