+++ src/app/components/ServiceWorkerRegister/hooks/useServiceWorkerRegister.ts
@@ -1,0 +1,4 @@
+/** Example hook for ServiceWorkerRegister */
+export const useServiceWorkerRegister = () => {
+  return {}
+}